# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CraftLand Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /creators:
    get:
      operationId: listCreators
      tags: [creators]
      summary: List all creators
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of creators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Creator"
    post:
      operationId: createCreator
      tags: [creators]
      summary: Create a new creator (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreatorBody"
      responses:
        "201":
          description: Created creator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
  /creators/{id}:
    get:
      operationId: getCreator
      tags: [creators]
      summary: Get creator profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Creator profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
        "404":
          description: Creator not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCreator
      tags: [creators]
      summary: Update a creator (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCreatorBody"
      responses:
        "200":
          description: Updated creator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
    delete:
      operationId: deleteCreator
      tags: [creators]
      summary: Delete a creator (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Creator deleted
  /creators/{id}/maps:
    get:
      operationId: getCreatorMaps
      tags: [creators]
      summary: Get all maps by a creator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Creator maps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Map"
  /creators/{id}/follow:
    post:
      operationId: followCreator
      tags: [creators]
      summary: Follow or unfollow a creator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Follow toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResponse"

  # Submissions
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Map:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        image:
          type: ["string", "null"]
        mapLink:
          type: ["string", "null"]
        creatorId:
          type: ["integer", "null"]
        creatorName:
          type: ["string", "null"]
        creatorLogo:
          type: ["string", "null"]
        likes:
          type: integer
        views:
          type: integer
        region:
          type: string
        isFeatured:
          type: boolean
        isTrending:
          type: boolean
        createdAt:
          type: string
        isLiked:
          type: boolean
        isBookmarked:
          type: boolean
      required:
        - id
        - name
        - code
        - likes
        - views
        - region
        - isFeatured
        - isTrending
        - createdAt
        - isLiked
        - isBookmarked
...
    FollowResponse:
      type: object
      properties:
        following:
          type: boolean
      required:
        - following
    Creator:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        logo:
          type: ["string", "null"]
        banner:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
        totalMaps:
          type: integer
        totalLikes:
          type: integer
        isFollowed:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - totalMaps
        - totalLikes
        - isFollowed
        - createdAt
    CreateCreatorBody:
      type: object
      properties:
        name:
          type: string
        logo:
          type: ["string", "null"]
        banner:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
      required:
        - name
    UpdateCreatorBody:
      type: object
      properties:
        name:
          type: string
        logo:
          type: ["string", "null"]
        banner:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        type:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        link:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - type
        - title
        - message
        - isRead
        - createdAt
...